byte_unit

Enum Unit

pub enum Unit {

    Bit,
    B,
    Kbit,
    Kibit,
    KB,
    KiB,
    Mbit,
    Mibit,
    MB,
    MiB,
    Gbit,
    Gibit,
    GB,
    GiB,
    Tbit,
    Tibit,
    TB,
    TiB,
    Pbit,
    Pibit,
    PB,
    PiB,
    Ebit,
    Eibit,
    EB,
    EiB,

}

The unit of bits/bytes.

Variants

Bit

8 Bit = 1 byte.

B

1 B = 1 byte.

Kbit

1 Kbit = 125 bytes.

Kibit

1 Kibit = 27 bytes.

KB

1 KB = 103 bytes.

KiB

1 KiB = 210 bytes.

Mbit

1 Mbit = 125 * 103 bytes.

Mibit

1 Mibit = 217 bytes.

MB

1 MB = 106 bytes.

MiB

1 MiB = 220 bytes.

Gbit

1 Gbit = 125 * 106 bytes.

Gibit

1 Gibit = 227 bytes.

GB

1 GB = 109 bytes.

GiB

1 GiB = 230 bytes.

Tbit

1 Tbit = 125 * 109 bytes.

Tibit

1 Tibit = 237 bytes.

TB

1 TB = 1012 bytes.

TiB

1 TiB = 240 bytes.

Pbit

1 Pbit = 125 * 1012 bytes.

Pibit

1 Pibit = 247 bytes.

PB

1 PB = 1015 bytes.

PiB

1 PiB = 250 bytes.

Ebit

1 Ebit = 125 * 1015 bytes.

Eibit

1 Eibit = 257 bytes.

EB

1 EB = 1018 bytes.

EiB

1 EiB = 260 bytes.

Implementations

impl Unit

Associated functions for parsing strings.

pub fn parse_str<S: AsRef<str>>( s: S, ignore_case: bool, prefer_byte: bool, ) -> Result<Self, UnitParseError>

Create a new Unit instance from a string. The string may be "", "B", "M", "MB", "MiB", "b", "Mb", "Mbit".

You can ignore the case of โBโ (byte), which means b will still be treated as bytes instead of bits.

If the input string is empty, it will return B if prefer_byte is true; otherwise, it will return b. Similarly, if the string is not empty but it does not explicitly contains "B", "b", or "bit", it will imply the base is "B" if prefer_byte is true; otherwise, imply the base is "b".

Examples

let unit = Unit::parse_str("Kib", true, true).unwrap(); // KiB

let unit = Unit::parse_str("Kib", false, true).unwrap(); // Kibit

impl Unit

Methods for converting a Unit instance into a primitive integer.

pub const fn as_bits_u128(self) -> u128

Retrieve the bit represented by this Unit instance.

Examples

use byte_unit::Unit;

assert_eq!(1, Unit::Bit.as_bits_u128());
assert_eq!(8, Unit::B.as_bits_u128());
assert_eq!(8000, Unit::KB.as_bits_u128());
assert_eq!(1024, Unit::Kibit.as_bits_u128());

impl Unit

Methods for converting a Unit instance into a string.

pub const fn as_str(self) -> &'static str

Retrieve the string represented by this Unit instance.

Examples

use byte_unit::Unit;

assert_eq!("B", Unit::B.as_str());
assert_eq!("KB", Unit::KB.as_str());
assert_eq!("MiB", Unit::MiB.as_str());
assert_eq!("Gb", Unit::Gbit.as_str());
assert_eq!("Tib", Unit::Tibit.as_str());

impl Unit

Methods for categorizing variants.

pub const fn is_bit(self) -> bool

Check whether the unit is based on bits.

Examples

use byte_unit::Unit;

assert_eq!(false, Unit::KB.is_bit());
assert_eq!(true, Unit::Kbit.is_bit());

pub const fn is_binary_multiples(self) -> bool

Check whether the unit is based on powers of 2.

Examples

use byte_unit::Unit;

assert_eq!(false, Unit::KB.is_binary_multiples());
assert_eq!(true, Unit::KiB.is_binary_multiples());