byte_unit

Struct AdjustedByte

pub struct AdjustedByte { /* private fields */ }

Generated from the Byte::get_adjusted_unit method or the the Byte::get_appropriate_unit method.

For accuracy representation, utilize the Byte struct.

Implementations

impl AdjustedByte

Methods for getting values.

pub const fn get_value(&self) -> f64

Get the value.

pub const fn get_unit(&self) -> Unit

Get the unit.

pub fn get_byte(&self) -> Byte

Create a new Byte instance from this AdjustedByte instance.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64_with_unit(1555, Unit::KB).unwrap();

let adjusted_byte = byte.get_adjusted_unit(Unit::MB);

let byte_back = adjusted_byte.get_byte();

assert_eq!(byte, byte_back);

Points to Note

The result may not be logically equal to the original Byte instance due to the accuracy of floating-point numbers.