byte_unit

Struct Byte

pub struct Byte(/* private fields */);

Representing the size in bytes.

Implementations

impl Byte

Associated functions for generating AdjustedByte.

pub fn get_adjusted_unit(self, unit: Unit) -> AdjustedByte

Adjust the unit and value for this Byte instance.

Examples

use byte_unit::{AdjustedByte, Byte, Unit};

let byte = Byte::parse_str("123KiB", true).unwrap();

let adjusted_byte = byte.get_adjusted_unit(Unit::KB);

assert_eq!("125.952 KB", adjusted_byte.to_string());

use byte_unit::{AdjustedByte, Byte, Unit};

let byte = Byte::parse_str("50.84 MB", true).unwrap();

let adjusted_byte = byte.get_adjusted_unit(Unit::MiB);

assert_eq!("48.48480224609375 MiB", adjusted_byte.to_string());

pub fn get_appropriate_unit(&self, unit_type: UnitType) -> AdjustedByte

Find the appropriate unit and value for this Byte instance.

Examples

use byte_unit::{Byte, UnitType};

let byte = Byte::parse_str("123KiB", true).unwrap();

let adjusted_byte = byte.get_appropriate_unit(UnitType::Decimal);

assert_eq!("125.952 KB", adjusted_byte.to_string());

use byte_unit::{Byte, UnitType};

let byte = Byte::parse_str("50.84 MB", true).unwrap();

let adjusted_byte = byte.get_appropriate_unit(UnitType::Binary);

assert_eq!("48.48480224609375 MiB", adjusted_byte.to_string());

impl Byte

Constant Bytes.

pub const BYTE: Byte

One byte.

pub const KILOBYTE: Byte

1 KB = 103 bytes.

pub const MEGABYTE: Byte

1 MB = 106 bytes.

pub const GIGABYTE: Byte

1 GB = 109 bytes.

pub const TERABYTE: Byte

1 TB = 1012 bytes.

pub const PETABYTE: Byte

1 PB = 1015 bytes.

pub const EXABYTE: Byte

1 EB = 1018 bytes.

pub const KIBIBYTE: Byte

1 KiB = 210 bytes.

pub const MEBIBYTE: Byte

1 MiB = 220 bytes.

pub const GIBIBYTE: Byte

1 GiB = 230 bytes.

pub const TEBIBYTE: Byte

1 TiB = 240 bytes.

pub const PEBIBYTE: Byte

1 PiB = 250 bytes.

pub const EXBIBYTE: Byte

1 EiB = 260 bytes.

pub const KILOBIT: Byte

1 Kbit = 125 bytes.

pub const MEGABIT: Byte

1 Mbit = 125 * 103 bytes.

pub const GIGABIT: Byte

1 Gbit = 125 * 106 bytes.

pub const TERABIT: Byte

1 Tbit = 125 * 109 bytes.

pub const PETABIT: Byte

1 Pbit = 125 * 1012 bytes.

pub const EXABIT: Byte

1 Ebit = 125 * 1015 bytes.

pub const KIBIBIT: Byte

1 Kibit = 27 bytes.

pub const MEBIBIT: Byte

1 Mibit = 217 bytes.

pub const GIBIBIT: Byte

1 Gibit = 227 bytes.

pub const TEBIBIT: Byte

1 Tibit = 237 bytes.

pub const PEBIBIT: Byte

1 Pibit = 247 bytes.

pub const EXBIBIT: Byte

1 Eibit = 257 bytes.

pub const MIN: Byte

0 byte.

pub const MAX: Byte

1027 - 1 bytes if the u128 feature is enabled, or 264 - 1 otherwise.

impl Byte

Associated functions for building Byte instances using Decimal.

pub fn from_decimal(size: Decimal) -> Option<Self>

Create a new Byte instance from a size in bytes.

Examples

use byte_unit::Byte;
use rust_decimal::Decimal;

let byte = Byte::from_decimal(Decimal::from(15000000u64)).unwrap(); // 15 MB

Points to Note

If the input size is too large (the maximum is 1027 - 1 if the u128 feature is enabled, or 264 - 1 otherwise) or not greater than or equal to 0, this function will return None.

The fractional part will be rounded up.

impl Byte

Associated functions for building Byte instances using Decimal (with Unit).

pub fn from_decimal_with_unit(size: Decimal, unit: Unit) -> Option<Self>

Create a new Byte instance from a size of bytes with a unit.

Examples

use byte_unit::{Byte, Unit};
use rust_decimal::Decimal;

let byte = Byte::from_decimal_with_unit(Decimal::from(15u64), Unit::MB).unwrap(); // 15 MB

Points to Note

If the calculated byte is too large or not greater than or equal to 0, this function will return None.

The calculated byte will be rounded up.

impl Byte

Methods for finding an unit using Decimal.

pub fn get_recoverable_unit( self, allow_in_bits: bool, precision: usize, ) -> (Decimal, Unit)

Find the appropriate unit and value that can be used to recover back to this Byte precisely.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64(3670016);

assert_eq!(
    (3.5f64.try_into().unwrap(), Unit::MiB),
    byte.get_recoverable_unit(false, 3)
);

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64(437500);

assert_eq!(
    (3.5f64.try_into().unwrap(), Unit::Mbit),
    byte.get_recoverable_unit(true, 3)
);

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64(437500);

assert_eq!(
    (437.5f64.try_into().unwrap(), Unit::KB),
    byte.get_recoverable_unit(false, 3)
);

Points to Note

precision should be smaller or equal to 26 if the u128 feature is enabled, otherwise 19. The typical precision is 3.

impl Byte

Associated functions for parsing strings.

pub fn parse_str<S: AsRef<str>>( s: S, ignore_case: bool, ) -> Result<Self, ParseError>

Create a new Byte instance from a string. The string may be "10", "10B", "10M", "10MB", "10MiB", "80b", "80Mb", "80Mbit".

You can ignore the case of “B” (byte), which means b will still be treated as bytes instead of bits.

Examples

let byte = Byte::parse_str("123Kib", true).unwrap(); // 123 * 1024 bytes

let byte = Byte::parse_str("123Kib", false).unwrap(); // 123 * 1024 bits = 123 * 1024 / 8 bytes

impl Byte

Associated functions for building Byte instances.

pub const fn from_u128(size: u128) -> Option<Self>

Create a new Byte instance from a size in bytes.

Examples

let byte = Byte::from_u128(15000000).unwrap(); // 15 MB

Points to Note

If the input size is too large (the maximum is 1027 - 1 if the u128 feature is enabled, or 264 - 1 otherwise), this function will return None.

pub const unsafe fn from_u128_unsafe(size: u128) -> Self

Create a new Byte instance from a size in bytes.

Examples

let byte = unsafe { Byte::from_u128_unsafe(15000000) }; // 15 MB

Safety

You must ensure the input size is not too large (the maximum is 1027 - 1 if the u128 feature is enabled, or 264 - 1 otherwise) on your own.

pub const fn from_u64(size: u64) -> Self

Create a new Byte instance from a size in bytes.

Examples

let byte = Byte::from_u64(15000000); // 15 MB

pub fn from_f64(size: f64) -> Option<Self>

Create a new Byte instance from a size in bytes.

Examples

let byte = Byte::from_f64(15000000.0).unwrap(); // 15 MB

Points to Note

If the input size is too large (the maximum is 1027 - 1 if the u128 feature is enabled, or 264 - 1 otherwise) or not greater than or equal to 0, this function will return None.

The fractional part will be rounded up.

pub fn from_f32(size: f32) -> Option<Self>

Create a new Byte instance from a size in bytes.

Examples

let byte = Byte::from_f32(15000000.0).unwrap(); // 15 MB

Points to Note

If the input size is too large (the maximum is 1027 - 1 if the u128 feature is enabled, or 264 - 1 otherwise) or not greater than or equal to 0, this function will return None.

The fractional part will be rounded up.

pub const fn from_i128(size: i128) -> Option<Self>

Create a new Byte instance from a size in bytes.

Examples

let byte = Byte::from_i128(15000000).unwrap(); // 15 MB

Points to Note

If the input size is too large (the maximum is 1027 - 1 if the u128 feature is enabled, or 264 - 1 otherwise) or negative, this function will return None.

pub const fn from_i64(size: i64) -> Option<Self>

Create a new Byte instance from a size in bytes.

Examples

let byte = Byte::from_i64(15000000).unwrap(); // 15 MB

Points to Note

If the input size is negative, this function will return None.

impl Byte

Associated functions for building Byte instances (with Unit).

pub const fn from_u128_with_unit(size: u128, unit: Unit) -> Option<Self>

Create a new Byte instance from a size of bytes with a unit.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_u128_with_unit(15, Unit::MB).unwrap(); // 15 MB

Points to Note

If the calculated byte is too large, this function will return None.

If the input unit is Bit, the calculated byte will be rounded up.

pub const fn from_u64_with_unit(size: u64, unit: Unit) -> Option<Self>

Create a new Byte instance from a size of bytes with a unit.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64_with_unit(15, Unit::MB).unwrap(); // 15 MB

Points to Note

If the calculated byte is too large, this function will return None.

If the input unit is Bit, the calculated byte will be rounded up.

pub fn from_f64_with_unit(size: f64, unit: Unit) -> Option<Self>

Create a new Byte instance from a size of bytes with a unit.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_f64_with_unit(15.0, Unit::MB).unwrap(); // 15 MB

Points to Note

If the calculated byte is too large or not greater than or equal to 0, this function will return None.

The calculated byte will be rounded up.

pub fn from_f32_with_unit(size: f32, unit: Unit) -> Option<Self>

Create a new Byte instance from a size of bytes with a unit.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_f32_with_unit(15.0, Unit::MB).unwrap(); // 15 MB

Points to Note

If the calculated byte is too large or not greater than or equal to 0, this function will return None.

The calculated byte will be rounded up.

pub const fn from_i128_with_unit(size: i128, unit: Unit) -> Option<Self>

Create a new Byte instance from a size of bytes with a unit.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_i128_with_unit(15, Unit::MB).unwrap(); // 15 MB

Points to Note

If the calculated byte is too large or negative, this function will return None.

The calculated byte will be rounded up.

pub const fn from_i64_with_unit(size: i64, unit: Unit) -> Option<Self>

Create a new Byte instance from a size of bytes with a unit.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_i64_with_unit(15, Unit::MB).unwrap(); // 15 MB

Points to Note

If the calculated byte is too large or negative, this function will return None.

The calculated byte will be rounded up.

impl Byte

Methods for converting a Byte instance into a primitive integer.

pub const fn as_u128(self) -> u128

Retrieve the byte represented by this Byte instance.

Examples

use byte_unit::Byte;

let byte = Byte::parse_str("123Kib", true).unwrap();

let result = byte.as_u128();

assert_eq!(125952, result);

use byte_unit::Byte;

let byte = Byte::parse_str("123Kib", false).unwrap();

let result = byte.as_u128();

assert_eq!(15744, result);

pub const fn as_u64(self) -> u64

Retrieve the byte represented by this Byte instance. When the u128 feature is enabled, if the byte is actually greater than 264 - 1, it will return 264 - 1.

Examples

use byte_unit::Byte;

let byte = Byte::parse_str("1kb", true).unwrap();

let result = byte.as_u64();

assert_eq!(1000, result);

use byte_unit::Byte;

let byte = Byte::parse_str("1zb", true).unwrap();

let result = byte.as_u64();

assert_eq!(u64::MAX, result);

pub const fn as_u64_checked(self) -> Option<u64>

Retrieve the byte represented by this Byte instance.

Examples

use byte_unit::Byte;

let byte = Byte::parse_str("1kb", true).unwrap();

let result = byte.as_u64_checked();

assert_eq!(Some(1000), result);

use byte_unit::Byte;

let byte = Byte::parse_str("1zb", true).unwrap();

let result = byte.as_u64_checked();

assert_eq!(None, result);

impl Byte

Methods for calculation.

pub const fn add(self, rhs: Byte) -> Option<Byte>

Add another Byte instance.

Examples

use byte_unit::Byte;

let byte_1 = Byte::from_u64(1024);
let byte_2 = Byte::from_u64(512);

let byte = byte_1.add(byte_2).unwrap();

assert_eq!(1536, byte.as_u64());

Points to Note

If the calculated byte is too large, this function will return None.

pub const fn subtract(self, rhs: Byte) -> Option<Byte>

Subtract another Byte instance.

Examples

use byte_unit::Byte;

let byte_1 = Byte::from_u64(1024);
let byte_2 = Byte::from_u64(512);

let byte = byte_1.subtract(byte_2).unwrap();

assert_eq!(512, byte.as_u64());

Points to Note

If the right-hand side is bigger then this Byte instance, this function will return None.

pub const fn multiply(self, rhs: usize) -> Option<Byte>

Multiplied by an unsigned integer.

Examples

use byte_unit::Byte;

let count = 100;
let byte = Byte::from_u64(1024);

let total_byte = byte.multiply(100).unwrap();

assert_eq!(102400, total_byte.as_u64());

Points to Note

If the calculated byte is too large, this function will return None.

pub const fn divide(self, rhs: usize) -> Option<Byte>

Divided by an unsigned integer.

Examples

use byte_unit::Byte;

let count = 100;
let byte = Byte::from_u64(1024);

let total_byte = byte.divide(100).unwrap();

assert_eq!(10, total_byte.as_u64());

Points to Note

If the input right-hand side is zero, this function will return None.

The result will be rounded down.

impl Byte

Methods for finding an unit.

pub const fn get_exact_unit(self, allow_in_bits: bool) -> (u128, Unit)

Obtain the largest unit which is the greatest factor of this Byte instance.

Examples

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64(3145728);

let (n, unit) = byte.get_exact_unit(true);

assert_eq!(3, n);
assert_eq!(Unit::MiB, unit);

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64(375000);

let (n, unit) = byte.get_exact_unit(true);

assert_eq!(3, n);
assert_eq!(Unit::Mbit, unit);

use byte_unit::{Byte, Unit};

let byte = Byte::from_u64(375000);

let (n, unit) = byte.get_exact_unit(false);

assert_eq!(375, n);
assert_eq!(Unit::KB, unit);